+++ nw/VIEW/ViewFigureLibraryState.cpp
@@ -1,0 +1,50 @@
+#define LAST_H__VIEW
+#include "game.h"
+
+#include <new>
+
+#include "ViewFigureLibraryState.h"
+
+namespace {
+
+bool g_figureLibraryReady = false;
+
+}  // namespace
+
+dword CViewTexture::m_dwDefaultInfoFlags = 0;
+
+CViewTexture::~CViewTexture() {
+  if (m_hImage != nullptr) GRDeleteTextureFromDB(m_hImage);
+}
+
+CViewLibTexturedFigure::SLibTexture*
+    CViewLibTexturedFigure::m_pLibTextures = nullptr;
+int CViewLibTexturedFigure::m_nLibTextures = 0;
+int CViewLibTexturedFigure::m_nMaxLibTextures = 0;
+
+void CViewLibTexturedFigure::LoadLib(TCchar* name) {
+  if (name != nullptr) return;
+  ClearLib();
+  m_pLibTextures = new (std::nothrow) SLibTexture[100];
+  m_nMaxLibTextures = m_pLibTextures != nullptr ? 100 : 0;
+  g_figureLibraryReady = m_pLibTextures != nullptr;
+}
+
+void CViewLibTexturedFigure::ClearLib() {
+  delete[] m_pLibTextures;
+  m_pLibTextures = nullptr;
+  m_nLibTextures = 0;
+  m_nMaxLibTextures = 0;
+}
+
+bool ViewFigureLibrary_Initialize() {
+  CViewLibTexturedFigure::LoadLib(nullptr);
+  return g_figureLibraryReady;
+}
+
+void ViewFigureLibrary_Release() {
+  CViewLibTexturedFigure::ClearLib();
+  g_figureLibraryReady = false;
+}
+
+bool ViewFigureLibrary_IsReady() { return g_figureLibraryReady; }